Day 5: part 1 complete

+++ day-5/src/main.rs
@@ -12,29 +12,29 @@ type ProcessMap = HashMap<u8, Vec<RangeData>>;
 
 #[derive(Debug)]
 struct RangeData {
-    min: u32,
-    max: u32,
-    offset: i32,
+    min: u64,
+    max: u64,
+    offset: i64,
 }
 
 impl RangeData {
-    pub fn new(destination_start: u32, source_start: u32, range_length: u32) -> Self {
+    pub fn new(destination_start: u64, source_start: u64, range_length: u64) -> Self {
         RangeData {
             min: source_start,
             max: source_start + range_length - 1,
-            offset: destination_start as i32 - source_start as i32,
+            offset: destination_start as i64 - source_start as i64,
         }
     }
 
-    // pub fn destination(self, source: u32) -> u32 {
-        // (source as i32 + self.offset) as u32
+    // pub fn destination(self, source: u64) -> u64 {
+        // (source as i64 + self.offset) as u64
      // }
 }
 
 fn main() {
     let input = fs::read_to_string("assets/input.txt").unwrap();
     let (seeds, process_map) = process_input(&input);
-    let mut min_location = u32::MAX;
+    let mut min_location = u64::MAX;
     for seed in seeds {
         let mut source = seed;
         for process_index in 0..7 {
@@ -46,10 +46,10 @@ fn main() {
     println!("{:?}", min_location);
 }
 
-fn process_input(input: &str) -> (Vec<u32>, ProcessMap) {
+fn process_input(input: &str) -> (Vec<u64>, ProcessMap) {
     let mut process_map: HashMap<u8, Vec<RangeData>> = HashMap::new();
     let mut current_index: i8 = -1;
-    let mut seeds: Vec<u32> = vec![];
+    let mut seeds: Vec<u64> = vec![];
 
     for line in input.lines() {
         if line == "" { 
@@ -72,27 +72,27 @@ fn process_input(input: &str) -> (Vec<u32>, ProcessMap) {
     (seeds, process_map)
 }
 
-fn get_numbers_from_line(line: &str) -> Vec<u32> {
-    let mut numbers: Vec<u32> = vec![];
+fn get_numbers_from_line(line: &str) -> Vec<u64> {
+    let mut numbers: Vec<u64> = vec![];
     let mut num_buffer = String::from("");
     for char in line.chars() {
         if char.is_digit(10) {
             num_buffer += &char.to_string();
         } else if !num_buffer.is_empty() && char == ' ' {
-            numbers.push(num_buffer.parse::<u32>().unwrap());
+            numbers.push(num_buffer.parse::<u64>().unwrap());
             num_buffer.clear();
         }
     }
-    numbers.push(num_buffer.parse::<u32>().unwrap());
+    numbers.push(num_buffer.parse::<u64>().unwrap());
     numbers
 }
 
-fn get_destination(process_map: &ProcessMap, process_index: u8, source: u32) -> u32 {
+fn get_destination(process_map: &ProcessMap, process_index: u8, source: u64) -> u64 {
     let ranges = process_map.get(&process_index).unwrap();
     for range in ranges {
         if  source >= range.min && source <= range.max {
             // TODO: use destination fn for this
-            return (source as i32 + range.offset) as u32;
+            return (source as i64 + range.offset) as u64;
         }
     }
     return source;
